requirements.txt
pushover-complete>=2.0.0
aiohttp>=3.9.0
playwright>=1.40.0
python-dotenv
